requirements.txt
Django>=3.2
requests>=2.25
python-dotenv
